chore(errors): add NotFound and create sep. file
