+++ src/routes.rs
@@ -1,55 +1,12 @@
 use sha2::{Digest, Sha256};
 use warp::{Rejection, Reply};
 
-use crate::{auth, models::User};
-
-#[derive(Debug)]
-struct Conflict;
-
-impl warp::reject::Reject for Conflict {}
-
-#[derive(Debug)]
-struct InternalError;
-
-impl warp::reject::Reject for InternalError {}
-
-#[derive(Debug)]
-struct NotAuthorized;
-
-impl warp::reject::Reject for NotAuthorized {}
-
-pub async fn handle_rejection(err: Rejection) -> Result<impl Reply, std::convert::Infallible> {
-    if err.is_not_found() {
-        Ok(crate::warp_reply!(
-            "There is no user with that name".to_owned(),
-            NOT_FOUND
-        ))
-    } else if let Some(e) = err.find::<auth::AuthorizationError>() {
-        Ok(crate::warp_reply!(format!("{:?}", e), BAD_REQUEST))
-    } else if err.find::<NotAuthorized>().is_some() {
-        Ok(crate::warp_reply!(
-            "The password is incorrect".to_string(),
-            UNAUTHORIZED
-        ))
-    }
-    else if err.find::<Conflict>().is_some() {
-        Ok(crate::warp_reply!(
-            "The provided username already exists".to_string(),
-            CONFLICT
-        ))
-    } else {
-        eprintln!("An unknown error occured: {:?}", err);
-        Ok(crate::warp_reply!(
-            "Something went wrong, please try again later.".to_owned(),
-            INTERNAL_SERVER_ERROR
-        ))
-    }
-}
+use crate::{auth, errors::*, models::User};
 
 pub async fn auth(
     db_pool: deadpool_postgres::Pool,
     auth_header: String,
-) -> std::result::Result<(deadpool_postgres::Pool, User), Rejection> {
+) -> Result<(deadpool_postgres::Pool, User), Rejection> {
     let auth::Auth { username, password } = auth::validate_header(auth_header.as_str())?;
     let pool = db_pool
         .get()
@@ -59,7 +16,7 @@ pub async fn auth(
     let user = crate::db::get_user(&pool, &username)
         .await
         .map_err(|err| match err {
-            tokio_pg_mapper::Error::ColumnNotFound => warp::reject::not_found(),
+            tokio_pg_mapper::Error::ColumnNotFound => warp::reject::custom(NotFound),
             _ => warp::reject::custom(InternalError),
         })?;
 
@@ -90,7 +47,7 @@ pub async fn create_account(
     let user = crate::db::get_user(&pool, &username)
         .await
         .map_err(|err| match err {
-            tokio_pg_mapper::Error::ColumnNotFound => warp::reject::not_found(),
+            tokio_pg_mapper::Error::ColumnNotFound => warp::reject::custom(NotFound),
             _ => warp::reject::custom(InternalError),
         });
 
@@ -117,5 +74,5 @@ pub async fn create_account(
 pub async fn get_account(
     (_, user): (deadpool_postgres::Pool, User),
 ) -> Result<impl Reply, Rejection> {
-    Ok(crate::warp_reply!(format!("Welcome back, {}!", user.username), CREATED))
+    Ok(crate::warp_reply!(format!("Welcome back, {}!", user.username), OK))
 }
